refactor: Grid UI 분리(#88)
- children만 렌더링하는 컴포넌트로 변경
- 기존 Grid는 파일에 담긴 내용에 맞춰 PopularMovieGrid로 변경
